% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{a_fun}
\alias{a_fun}
\title{The R-DIF scaling function for item slopes.}
\usage{
a_fun(mle, log = F)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{log}{logical: return of log(a2/a1)?}
}
\value{
The vector of scaling function values.
}
\description{
Computes the scaling function \code{a2/a1} for item slopes (a) in groups g = \{1, 2\}
}
