% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{apply_bias}
\alias{apply_bias}
\title{Apply bias to a random subset of item intercepts}
\usage{
apply_bias(b0, n.biased = 1, max.bias = 0.5, worst.case = T)
}
\arguments{
\item{b0}{item difficult (not intercept) in reference group}

\item{n.biased}{number of items to bias}

\item{max.bias}{maximum amount of bias to apply}

\item{worst.case}{logical: apply worst case bias? Default is \code{TRUE}. If set to \code{FALSE}, bias is sampled from \code{runif(max.bias/2, max.bias)}}
}
\value{
A vector of item intercepts, with bias added
}
\description{
Apply bias to a random subset of item intercepts
}
