% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTF_functions.R
\name{bsq_var_weight}
\alias{bsq_var_weight}
\title{Weights used to computed the variance of the RDIF scaling parameter estimated via IRLS.}
\usage{
bsq_var_weight(mle, theta, par = "intercept", log = F, alpha = 0.05)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{theta}{the value of the scaling parameter (e.g., outputted from \code{\link[robustDIF]{rdif}})}

\item{par}{character: use the scaling function for item intercepts or item slopes? One of \code{c("intercept", "slope")}.}

\item{log}{logical: use log of scaling function? Only applies if \code{par = "slope"}.}

\item{alpha}{the desired false positive rate for flagging items with DIF.}
}
\value{
A vector of weights v used to compute variance of RDIF scale parameter
}
\description{
These are not the IRLS weights, w, but the weights used for the variance for the IRLS estimate, v = psi_prime / sum w.
}
\details{
If \code{abs(u) > k} , \code{psi_prime(u) = 0}. Else,
 \code{psi_prime(u) = (1 - (u/k)^2)^2 - (2u/k)^2 (1 - (u/k)^2)}.
}
