% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{bsq_weight}
\alias{bsq_weight}
\title{Computes the weights of the bi-square function for use with IRLS}
\usage{
bsq_weight(theta, y, var.y, alpha = 0.05)
}
\arguments{
\item{theta}{IRT scale parameter}

\item{y}{the output of \code{robustDIF::y_fun}}

\item{var.y}{the variance of y}

\item{alpha}{the desired false positive rate for flagging items with DIF}
}
\value{
bi-square weights for (y - theta) / var.y, for each item
}
\description{
Computes the weights of the bi-square function for use with IRLS
}
