% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{chi2_test}
\alias{chi2_test}
\title{R-DIF chi-square test.}
\usage{
chi2_test(theta.y, theta.z, irt.mle, log = F)
}
\arguments{
\item{theta.y}{The IRT scale parameter for the item intercepts (mu / sigma)}

\item{theta.z}{The IRT scale parameter for the item slopes (sigma or log sigma).}

\item{irt.mle}{The output of \code{robustDIF::irt_mle}}

\item{log}{Logical: Use log of scaling function for item slopes?}
}
\value{
A named list with containing the value of the chi2.test and p(chi.square > chi2.test), for each item.
}
\description{
Simultaneously tests  for DIF in both the item intercept or slopes using a asymptotic chi-square test.
}
