% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{d_fun}
\alias{d_fun}
\title{The R-DIF scaling functions for item intercepts / thresholds.}
\usage{
d_fun(mle, type = 3)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{type}{a number in \code{1:3} indicating which version of delta to compute. See description for details.}
}
\value{
The vector of scaling function values.
}
\description{
Computes the scaling function to be used from the item thresholds (d) in groups = \{1, 2\}. The options are:
 \itemize{
 \item \code{type = 1}: computes \code{d.fun1 = (d2 - d1)/a1}
 \item \code{type = 2}: computes \code{d.fun2 = (d2 - d1)/a2}
 \item \code{type = 3}: computes \code{d.fun3 = (d2 - d1)/sqrt{(a1^2 + a2^2)/2}}
 }
}
\details{
Computes \code{(d2 - d1)/a} for each threshold (d) of each item in groups g = \{1, 2\}. The parameter 'a' depends on the value of \code{type}:
}
