% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{d_test}
\alias{d_test}
\title{R-DIF test for item intercepts}
\usage{
d_test(theta, irt.mle, alpha = 0.05)
}
\arguments{
\item{theta}{IRT scale parameter}

\item{irt.mle}{the output of \code{robustDIF::irt_mle}}

\item{alpha}{the desired false positive rate for flagging items with DIF}
}
\value{
A n.items-dimensional vector of p-values
}
\description{
R-DIF test for item intercepts
}
