% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{decision_errors}
\alias{decision_errors}
\title{Post processing for \code{sim_study} to compute decision errors}
\usage{
decision_errors(ds)
}
\arguments{
\item{dif}{The DIF table output from  \code{sim_study}}
}
\value{
A data.frame with false and true positives for each method length
}
\description{
Post processing for \code{sim_study} to compute decision errors
}
