% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{dif_test}
\alias{dif_test}
\title{Wald tests of differential item functioning (DIF).}
\usage{
dif_test(object, theta = NULL, fun = "d_fun3")
}
\arguments{
\item{object}{either the output of \code{\link[robustDIF]{get_model_parms}} or an \code{rdif} object from \code{\link[robustDIF]{rdif}}.}

\item{theta}{the estimated scaling parameter from \code{\link[robustDIF]{rdif}}}

\item{fun}{one of \code{c("a_fun1", "a_fun2", "d_fun1", "d_fun2", "d_fun3")}. See description for details.}
}
\value{
A data.frame whose rows containing the results of the test for each item parameter.
}
\description{
A Wald test of DIF on each item. Called internally by \code{\link[robustDIF]{rdif}}
}
\examples{
\dontrun{
mod <- rdif(mle = rdif.eg)
dif_test(object = rdif.eg, theta = mod$est)
dif_test(mod)
}
}
