% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdif_class_functions.R
\name{.onLoad}
\alias{.onLoad}
\title{Registers S3 methods at load time:
- print for class "rdif"
- plot for class "rdif"
- summary for class "rdif"}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{pkgname}{}
}
\description{
Registers S3 methods at load time:
- print for class "rdif"
- plot for class "rdif"
- summary for class "rdif"
}
