% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_functions.R
\name{format_pars}
\alias{format_pars}
\title{Helper function used to format parameters estimates}
\usage{
format_pars(pars, names.vec, type)
}
\arguments{
\item{pars}{numeric vector of item parameter estimates}

\item{names.vec}{character vector item names}

\item{type}{character; are \code{pars} from \code{"lavaan"} or \code{"mplus"}?}
}
\value{
data.frame of item parameter estimates
}
\description{
Helper function used to format parameters estimates
}
\seealso{
[robustDIF::get_lavaan_params()], [robustDIF::get_mplus_params()]
}
