% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{get_irt_mle}
\alias{get_irt_mle}
\title{Extract and format 2PL MLEs and their asymptotic VCOV matrix}
\usage{
get_irt_mle(fit.list, type = "mirt")
}
\arguments{
\item{fit.list}{a list of model fits.}

\item{type}{a string indicating the package that produced the fits}
}
\value{
a named list of MLEs and VCOVs
}
\description{
Takes a list of 2PL model fits and formats the MLEs and their VCOV matrix for use with \code{robustDIF}. Current implementation only supports lists of length 2 (i.e., two groups) and the first fit is treated as the reference group. The only type of fit currently supported is the \code{SingleGroupClass} of the \code{mirt} package. To use fits from other software, the user can manually format the MLEs and VCOVs in the same format as this function.
}
