% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{get_irt_pars}
\alias{get_irt_pars}
\title{Extract and format 2PL item parameter estimates and their covariance matrix.}
\usage{
get_irt_pars(fit.list, type = "mirt")
}
\arguments{
\item{fit.list}{a list of 2PL model fits.}

\item{type}{character: the name of the package that produced the fits (the current implementation only supports \code{\link[mirt]{mirt}}'s \code{SingleGroupClass}).}
}
\value{
A named list of item parameter estimates and covariance matrices.
}
\description{
Takes a list of 2PL model fits and formats the item parameter estimates and their covariance matrix. All \code{robustDIF} functions assume that the estimates were obtained by maximum likelihood and the covariance is asymptotically correct.

Note that the only type of fit currently supported is the \code{SingleGroupClass} of the \code{\link[mirt]{mirt}} package. Also, the current implementation only supports lists of length 2 (i.e., two groups) and the first fit is treated as the reference group.

It is possible to use fits from other software with \code{robustDIF} functions, but the parameter estimates and their covariance matrices must be formatted in a particular way. For more details, see the documentation for the example dataset \code{\link[robustDIF]{rdif.eg}}.
}
\seealso{
\code{\link[robustDIF]{rdif.eg}}
}
