% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{get_mirt_mle}
\alias{get_mirt_mle}
\title{Extract and format 2PL item parms from mirt}
\usage{
get_mirt_mle(mirt.fit.2pl)
}
\arguments{
\item{mirt.fit.2pl}{a mirt object (SingleGroupClass) for the 2PL model}
}
\value{
a n.items X 2 data.frame of 2PL item parameter MLEs
}
\description{
Extract and format 2PL item parms from mirt
}
