% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{get_mirt_vcov}
\alias{get_mirt_vcov}
\title{Extract and format 2PL VCOV from mirt}
\usage{
get_mirt_vcov(mirt.fit.2pl)
}
\arguments{
\item{mirt.fit.2pl}{a mirt object (SingleGroupClass) for the 2PL model}
}
\value{
a n.items X n.items covariance matrix for item parms
}
\description{
Extract and format 2PL VCOV from mirt
}
