% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{get_mplus_params}
\alias{get_mplus_params}
\title{Extract item parameter estimates and their covariance matrix from MPlus.}
\usage{
get_mplus_params(mplus.object)
}
\arguments{
\item{mplus.object}{an object of class \code{mplus.model} from \code{\link[MplusAutomation]{readModels}}. Expected to be a 1-factor model. Mplus .out file must include Tech3.}
}
\value{
A three-element \code{list}:
\itemize{
\item vector of parameter names taking the form "item.param"
\item vector of item parameter estimates
\item covariance matrix of item parameter estimates
}
}
\description{
Extract item parameter estimates and their covariance matrix from MPlus.
}
