% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{get_starts}
\alias{get_starts}
\title{Computes the staring values for IRLS (or NR)}
\usage{
get_starts(irt.mle, drops = T)
}
\arguments{
\item{irt.mle}{the output of \code{robustDIF::get_irt_mle}}

\item{drops}{logical: omit items if y changes substantially when changing ref group?}
}
\value{
a vector with elements med(Y), LTS(Y), and min Rho
}
\description{
Computes the staring values for IRLS (or NR)
}
