% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{grad_d}
\alias{grad_d}
\title{The gradient matrix of \code{\link[robustDIF]{d_fun}}.}
\usage{
grad_d(mle, theta = NULL, type = 3)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{theta}{(optional) the scaling parameter. Replaces item-specific values of d_fun if provided.}

\item{type}{a number in \code{1:3} indicating which version of delta to compute. See description for details.}
}
\value{
A matrix in which the columns are the gradient vectors of \code{\link[robustDIF]{d_fun}}, for each item and threshold.
}
\description{
The gradient is taken with respect to the item parameters and organized to be conformable with \code{Matrix::bdiag(mle$var.cov)}. When evaluating the gradient under the null hypothesis of no DIF, the optional argument \code{theta} can be provided. It replaces the item-specific values of d_fun in the gradient computation.
}
\seealso{
\code{\link[robustDIF]{d_fun}}
}
