% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{grad_mat}
\alias{grad_mat}
\title{Helper function to put gradient vectors into a matrix.}
\usage{
grad_mat(grad.vec)
}
\arguments{
\item{grad.vec}{A vector containing the nonzero elements of \code{\link[robustDIF]{grad_intercept}} or \code{\link[robustDIF]{grad_slope}}.}
}
\value{
A matrix in which the columns are the gradient vectors for each item.
}
\description{
Called internally by gradient functions.
}
