% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{grad_slope}
\alias{grad_slope}
\title{Compute the gradient of \code{y_slope}.}
\usage{
grad_slope(theta, irt.mle, log = F)
}
\arguments{
\item{theta}{The IRT scale parameter (sigma or log sigma).}

\item{irt.mle}{The output of \code{robustDIF::get_irt_mle}.}

\item{log}{Logical: Use log of scaling function?.}
}
\value{
A matrix in which the columns are the 4*n.item-dimensional gradient vectors of \code{y_slope} for each item.
}
\description{
For each item, the non-null elements of the gradient are organized as \code{c(a0, d0, a1, d1)}. The parameters a0 and d0 are the item slope and intercept in the reference group. The parameters a1 and d1 are the item slope and intercept in the comparison group.
}
