% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{grid}
\alias{grid}
\title{Performs grid search for minimum value of the bi-square Rho function}
\usage{
grid(irt.mle, alpha = 0.05, grid.width = 0.05, drops = T)
}
\arguments{
\item{irt.mle}{the output of \code{robustDIF::get_irt_mle}}

\item{alpha}{the desired false positive rate for flagging items with DIF}

\item{grid.width}{the width of grid points}

\item{drops}{passed from \code{get_starts}; leave at default value (\code{T}) to keep all items}
}
\value{
The location parameter that minimizes the bi-square Rho function
}
\description{
Performs grid search for minimum value of the bi-square Rho function
}
