% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{irls}
\alias{irls}
\title{IRLS for an IRT scale parameter using the bi-square function}
\usage{
irls(
  irt.mle,
  par = "intercept",
  log = F,
  alpha = 0.05,
  starting.value = "all",
  tol = 1e-07,
  maxit = 100
)
}
\arguments{
\item{irt.mle}{the output of \code{robustDIF::irt_mle}}

\item{par}{Use scaling function for item intercepts or item slopes? One of \code{c("intercept", "slope")}.}

\item{log}{Logical: Use log of scaling function? Only applies if \code{par = "slope"}.}

\item{alpha}{the desired false positive rate for flagging items with DIF}

\item{starting.value}{one of \code{c("med", "lts", "min_rho", "all")} or a numerical value to be used as the starting value. The default is \code{"all"} which returns the median of the other three.}

\item{tol}{convergence criterion for comparing subsequent values of estimate}

\item{maxit}{maximum number of iterations to perform}
}
\value{
A named list containing the estimate, weights, and number of iterations
}
\description{
IRLS for an IRT scale parameter using the bi-square function
}
