% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{joint_vcov}
\alias{joint_vcov}
\title{Helper function to merge two VCOV matrices.}
\usage{
joint_vcov(irt.mle)
}
\arguments{
\item{irt.mle}{The output of \code{robustDIF::get_irt_mle}.}
}
\value{
A block diagonal VCOV matrix in which each block is the VCOV of an item.
}
\description{
Puts two 2PL models into a single block diagonal VCOV matrix that is conformable with the output of the \code{robustDIF} gradient functions.
}
