% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{lasso}
\alias{lasso}
\title{Wrapper for GPMClasso}
\usage{
lasso(dat0, dat1)
}
\arguments{
\item{dat0}{item response data for reference group}

\item{dat1}{item response data for comparison group}
}
\value{
A named list with dif parameters for the item difficulties, and the estimated mean of .
}
\description{
Wrapper for GPMClasso
}
