% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{lr}
\alias{lr}
\title{Wrapper for mirt's implementation of the likelihood ratio (LR) test of DIF}
\usage{
lr(dat0, dat1, parms = c("d"))
}
\arguments{
\item{dat0}{item response data for reference group}

\item{dat1}{item response data for comparison group}

\item{parms}{which parameters to tests for DIF}
}
\value{
A named list with p values from the LR test and the scale parameters estimated in the model that assumes invariance
}
\description{
Wrapper for mirt's implementation of the likelihood ratio (LR) test of DIF
}
