% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{lr_chi}
\alias{lr_chi}
\title{Wrapper for mirt's implementation of the likelihood ratio (LR) test of DIF for item slopes and intercepts.}
\usage{
lr_chi(dat0, dat1)
}
\arguments{
\item{dat0}{item response data for reference group}

\item{dat1}{item response data for comparison group}
}
\value{
A named list with p values from the LR test and the scale parameters estimated in the model that assumes invariance
}
\description{
Wrapper for mirt's implementation of the likelihood ratio (LR) test of DIF for item slopes and intercepts.
}
