% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{lr_z}
\alias{lr_z}
\title{Wrapper for mirt's implementation of the likelihood ratio (LR) test of DIF for item intercept only}
\usage{
lr_z(dat0, dat1, which.par = "d")
}
\arguments{
\item{dat0}{item response data for reference group}

\item{dat1}{item response data for comparison group}
}
\value{
A named list with p values from the LR test and the scale parameters estimated in the model that assumes invariance
}
\description{
Wrapper for mirt's implementation of the likelihood ratio (LR) test of DIF for item intercept only
}
