% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{lts}
\alias{lts}
\title{Compute the least trimmed squares (LTS) estimate of location}
\usage{
lts(y, p = 0.5)
}
\arguments{
\item{y}{A vector of data points.}

\item{p}{The proportion of data points to trim.}
}
\value{
The LTS estimate of location.
}
\description{
Compute the least trimmed squares (LTS) estimate of location
}
