% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{mh}
\alias{mh}
\title{Wrapper for difR's implementation of the Mantel-Haenszel test of (uniform) DIF}
\usage{
mh(dat0, dat1)
}
\arguments{
\item{dat0}{item response data for reference group}

\item{dat1}{item response data for comparison group}
}
\value{
A named list with p values from the MH test, and the scale parameter estimated from the total scores
}
\description{
Wrapper for difR's implementation of the Mantel-Haenszel test of (uniform) DIF
}
