% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTF_functions.R
\name{ml_est}
\alias{ml_est}
\title{Maximum likelihood estimate (MLE) of scaling parameters.}
\usage{
ml_est(mle, par = "intercept", log = F)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{par}{character: use the scaling function for item intercepts or item slopes? One of \code{c("intercept", "slope")}.}

\item{log}{logical: use log of scaling function? Only applies if \code{par = "slope"}.}
}
\value{
The MLE of scaling parameter. Used internally to test DTF by comparing to RDIF estimate.
}
\description{
Maximum likelihood estimate (MLE) of scaling parameters.
}
