% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdif_class_functions.R
\name{plot.rdif}
\alias{plot.rdif}
\title{S3 plot method for objects of class "rdif"
Plots the rho function of the output from rdif.}
\usage{
\method{plot}{rdif}(object, ...)
}
\arguments{
\item{object}{A saved output from \code{rdif()} that is to be plotted.}

\item{...}{Additional arguments to be passed to \code{plot()}}
}
\value{
An object of class rdif, a plot with components:
\itemize {
\item{\code{xlab}}{theta}
\item{\code{ylab}}{rho}
\item{\code{type}}{line}
}
}
\description{
S3 plot method for objects of class "rdif"
Plots the rho function of the output from rdif.
}
\examples{
\dontrun{
# Assuming "rdif.eg" is a list of parameter values
rho <- rdif(mle = rdif.eg, fun = "d_fun3", grid.width = .01)
plot(rho)
}
}
