% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdif_class_functions.R
\name{print.rdif}
\alias{print.rdif}
\title{S3 print method for objects of class "rdif"
Prints the estimated scaling parameter from rdif. (to be added: SE)}
\usage{
\method{print}{rdif}(x, ...)
}
\arguments{
\item{x}{An object of class 'rdif', the saved output from the \code{rdif()} function.}

\item{...}{Additional arguments passed through from the generic.}
}
\value{
The input object, returned invisibly.
}
\description{
S3 print method for objects of class "rdif"
Prints the estimated scaling parameter from rdif. (to be added: SE)
}
\keyword{internal}
