% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\docType{data}
\name{rdif.eg}
\alias{rdif.eg}
\title{Example data set for R-DIF functions.}
\format{
A named list with 4 components:
\describe{
  \item{par0}{A \code{data.frame} or named \code{list} with \code{par0$a} containing the item slopes and \code{par0$d} containing the item intercepts, for the reference group.}
  \item{par1}{The item parameter estimates of the comparison groups. See \code{par0} for formatting.}
  \item{vcov0}{The covariance matrix of \code{par0}, formatted as either a \code{data.frame} or \code{matrix}. The parameters should be organized by item, with the slope parameter coming first and the intercept parameter coming second (e.g., \code{a.item1, d.item1, a.item2, d.item2, ...}).}
   \item{vcov1}{The covariance matrix of \code{par1}. See \code{vcov0} for formatting.}
}
}
\usage{
rdif.eg
}
\description{
A named list containing the maximum likelihood estimates and their estimated covariance matrix, for the 2PL IRT model fitted to 5 items in two independent groups. The first item has additive bias of .5 applied to the intercept only. The groups have a mean difference of .5 standard deviations on the latent trait. The variances of the latent trait are equal in each group.
}
\keyword{datasets}
