% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{rdif_z_test}
\alias{rdif_z_test}
\title{The R-DIF test of a single item parameter.}
\usage{
rdif_z_test(mle, par = "intercept", log = F)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{par}{character: use the scaling function for item intercepts or item slopes? One of \code{c("intercept", "slope")}.}

\item{log}{logical: use log of scaling function? Only applies if \code{par = "slope"}.}
}
\value{
A \code{data.frame} containing the value of the z.test and p(|z| > |z.test|), for each item parameter.
}
\description{
Tests for DIF in either the item intercepts / thresholds or slopes using the asymptotic z-test in Theorem 1 of Halpin 2022.
}
\examples{
# Test intercepts, using the built-in example dataset "rdif.eg"
\dontrun{
rdif_z_test(mle = rdif.eg)
}

# Test slopes
\dontrun{
rdif_z_test(.mle = rdif.eg, par = "slope")
}
}
