% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{rho_fun}
\alias{rho_fun}
\title{Computes the bi-square Rho function}
\usage{
rho_fun(irt.mle, par = "intercept", log = F, alpha = 0.05, grid.width = 0.05)
}
\arguments{
\item{irt.mle}{The output of \code{robustDIF::get_irt_mle}.}

\item{par}{Use scaling function for item intercepts or item slopes? One of \code{c("intercept", "slope")}.}

\item{log}{Logical: Use log of scaling function? Only applies if \code{par = "slope"}.}

\item{alpha}{The desired false positive rate for flagging items with DIF.}

\item{grid.width}{The width of grid points for \code{theta}.}

\item{theta}{The IRT scale parameter (mu / sigma for intercepts, sigma or log sigma for slopes).}
}
\value{
A named list with theta and Rho values, can be passed directly to plot.
}
\description{
Plotting the objective function of the minimization problem in theta. Useful for diagnosing local solutions.
}
