% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{rho_grid}
\alias{rho_grid}
\title{Grid search for minimum value of the bi-square Rho function.}
\usage{
rho_grid(y, var_fun, alpha = 0.05, grid.width = 0.05)
}
\arguments{
\item{y}{output of \code{\link[robustDIF]{get_starts}}}

\item{var_fun}{function that variance of \code{y} as a function of only the scale parameter.}

\item{alpha}{The desired false positive rate for flagging items with DIF.}

\item{grid.width}{The width of grid points.}
}
\value{
The location parameter that minimizes the bi-square Rho function.
}
\description{
This function is used by \code{\link[robustDIF]{get_starts}}, so its inputs are formatted for to take internal arguments of that function. For plotting, use \code{\link[robustDIF]{rho_fun}} instead.
}
\seealso{
\code{\link[robustDIF]{get_starts}}
}
