% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{sim_study}
\alias{sim_study}
\title{Runs a simulation study comparing R-DIF, the likelihood ratio test, the Mantel-Haenszel test, and the output of GPMClasso}
\usage{
sim_study(
  n.reps = 100,
  n.persons = 500,
  n.items = 15,
  n.biased = 0,
  bias = 0,
  impact = c(0, 1)
)
}
\arguments{
\item{n.reps}{number of replications}

\item{n.persons}{number of respondents per group}

\item{n.biased}{number of items with DIF on item intercepts}

\item{bias}{magnitude of bias (the same for all items with bias)}

\item{impact}{the mean and variance of the latent trait in the reference groups}

\item{n.item}{number of items}
}
\value{
A n.reps length list, where each element is a list containing (a) a data.frame with the output of the DIF procedures, and (b) a vector with estimates of impact from the procedures.
}
\description{
Runs a simulation study comparing R-DIF, the likelihood ratio test, the Mantel-Haenszel test, and the output of GPMClasso
}
