% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sim.R
\name{sim_study2}
\alias{sim_study2}
\title{Runs a simulation study comparing R-DIF and the likelihood ratio test}
\usage{
sim_study2(n.reps = 100, n.persons = 500, n.items = 15, bias = c(0.5, 1))
}
\arguments{
\item{n.reps}{number of replications}

\item{n.persons}{number of respondents per group}

\item{bias}{magnitude of bias (the same for all items with bias)}

\item{n.item}{number of items}

\item{n.biased}{number of items with DIF on item intercepts}

\item{impact}{the mean and variance of the latent trait in the reference groups}
}
\value{
A n.reps length list, where each element is a list containing (a) a data.frame with the output of the DIF procedures, and (b) a vector with estimates of impact from the procedures.
}
\description{
Runs a simulation study comparing R-DIF and the likelihood ratio test
}
