% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdif_class_functions.R
\name{summary.rdif}
\alias{summary.rdif}
\title{S3 summary method for objects of class "rdif"}
\usage{
\method{summary}{rdif}(object, ...)
}
\arguments{
\item{object}{An object of class 'rdif', a saved list of values from \code{rdif()}}

\item{...}{Additional arguments passed through from the generic.}
}
\value{
A printed summary of values
}
\description{
S3 summary method for objects of class "rdif"
}
\keyword{internal}
