% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{y_intercept}
\alias{y_intercept}
\title{The R-DIF scaling function for item intercepts / thresholds}
\usage{
y_intercept(mle)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}
}
\value{
A vector of Y values.
}
\description{
Computes the scaling function \code{Y = (d1 - d0)/a1} for each intercept / threshold of each item. Used to test DIF on item intercepts.
}
