% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{z_test}
\alias{z_test}
\title{R-DIF z-test.}
\usage{
z_test(theta, irt.mle, par = "intercept", log = F)
}
\arguments{
\item{theta}{The IRT scale parameter.}

\item{irt.mle}{The output of \code{robustDIF::irt_mle}.}

\item{par}{Use scaling function for item intercepts or item slopes? One of \code{c("intercept", "slope")}.}

\item{log}{Logical: Use log of scaling function? Only applies if \code{par = "slope"}.}
}
\value{
A named list with containing the value of the z.test and p(|z| > z.test), for each item.
}
\description{
Tests for DIF in either the item intercept or slopes using a asymptotic z-test.
}
