% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{bsq_weight}
\alias{bsq_weight}
\title{The bi-square weight function.}
\usage{
bsq_weight(u, k = 1.96)
}
\arguments{
\item{u}{Can be a single value, vector, or matrix.}

\item{k}{The tuning parameter. Can be a scalar or the same dimension as \code{u}.}
}
\value{
The bi-square psi-prime function.
}
\description{
If \code{abs(u) > k} , \code{psi(u) = 0}. Else, \code{psi(u) = (1 - (u/k)^2)^2}.
}
