% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{cov_yz}
\alias{cov_yz}
\title{Computes the asymptotic covariance of IRT scaling functions for the item intercepts and the item slopes, under the null hypothesis.}
\usage{
cov_yz(theta.y, theta.z, irt.mle, log = F)
}
\arguments{
\item{theta.y}{The IRT scale parameter for the item intercepts (mu / sigma)}

\item{theta.z}{The IRT scale parameter for the item slopes (sigma or log sigma).}

\item{irt.mle}{The output of \code{robustDIF::get_irt_mle}.}

\item{log}{Logical: Use log of scaling function for the slope?}
}
\value{
An n.item-dimensional vector that contains the covariance of the IRT scaling functions, for each item.
}
\description{
Used to compute the the quadratic form test of item intercepts and slopes.
}
