% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{cov_yz}
\alias{cov_yz}
\title{Compute the covariance of the asymptotic null distribution of IRT scaling functions.}
\usage{
cov_yz(theta.y, theta.z, irt.mle, log = F)
}
\arguments{
\item{theta.y}{the IRT scale parameter for the item intercepts}

\item{theta.z}{the IRT scale parameter for the item slopes}

\item{irt.mle}{the output of \code{\link[robustDIF]{get_irt_pars}}.}

\item{log}{logical: use of scaling function for the slopes?}
}
\value{
A vector that contains the covariance of the IRT scaling functions, for each item.
}
\description{
Compute the covariance of the asymptotic null distribution of IRT scaling functions.
}
\seealso{
\code{\link[robustDIF]{y_fun}}, \code{\link[robustDIF]{var_y}}
}
