% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTF_functions.R
\name{delta_test}
\alias{delta_test}
\title{Compares RDIF estimate of impact to maximum likelihood estimate.}
\usage{
delta_test(mle, par = "intercept", log = F, alpha = 0.05)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{par}{character: use the scaling function for item intercepts or item slopes? One of \code{c("intercept", "slope")}.}

\item{log}{logical: use log of scaling function? Only applies if \code{par = "slope"}.}

\item{alpha}{the desired false positive rate for flagging items with DIF.}
}
\value{
A named list with details of the test.
}
\description{
Asymptotic test of difference between RDIF estimate of impact and ML estimate of impact (based on IRT scaling functions).
}
