% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{delta_test}
\alias{delta_test}
\title{Wald tests of overall DIF in IRT scale parameters.}
\usage{
delta_test(mle, fun = "d_fun3", alpha = 0.05)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{fun}{one of \code{c("a_fun1", "a_fun2", "d_fun1", "d_fun2", "d_fun3")}. See description for details.}

\item{alpha}{the desired false positive rate for flagging items with DIF.}
}
\value{
A data.frame whose rows containing the results of the test for each variation of the bi-square loss function.
}
\description{
Tests for overall DIF in the IRT scale parameter using several variations of the bi-square loss function.
}
\examples{
# Test thresholds, using the built-in example dataset "rdif.eg"
\dontrun{delta_test(mle = rdif.eg, fun = "d_fun3")}
}
