% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{delta_test}
\alias{delta_test}
\title{Wald test of differential test functioning.}
\usage{
delta_test(mle, theta, k, fun = "d_fun3")
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{theta}{the estimated scaling parameter from \code{\link[robustDIF]{rdif}}}

\item{k}{the tuning parameter from \code{\link[robustDIF]{rdif}}}

\item{fun}{one of \code{c("a_fun1", "a_fun2", "d_fun1", "d_fun2", "d_fun3")}. See description for details.}
}
\value{
A data.frame that contains the output of the test.
}
\description{
A Wald test of the difference between the unweighted mean of the \code{\link[robustDIF]{y_fun}} and robust scaling parameter from \code{\link[robustDIF]{rdif}}. Called internally by \code{\link[robustDIF]{rdif}}
}
\examples{
#
\dontrun{
mod <- rdif(mle = rdif.eg)
delta_test(mle = rdif.eg, theta = mod$est)
}
}
