% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{delta_test_from_dif}
\alias{delta_test_from_dif}
\title{Wald test of differential test functioning.}
\usage{
delta_test_from_dif(mle, dif.items, fun = "d_fun3")
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{dif.items}{the indices of the items with DIF.}

\item{fun}{one of \code{c("a_fun1", "a_fun2", "d_fun1", "d_fun2", "d_fun3")}. See description for details.}
}
\value{
A data.frame that contains the output of the test.
}
\description{
A Wald test of the difference between the unweighted mean of the \code{\link[robustDIF]{y_fun}} computed for all items, and the unweighted mean excluding \code{dif.items}.
}
\examples{
#
\dontrun{
# Test for DTF omitting the first two items.
delta_test_from_dif(mle = rdif.eg, dif.items = c(1, 2))
}
}
