% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{get_mirt_params}
\alias{get_mirt_params}
\title{Extract item parameter estimates and their covariance matrix from \code{\link[mirt]{mirt}}.}
\usage{
get_mirt_params(mirt.object)
}
\arguments{
\item{mirt.object}{a \code{\link[mirt]{mirt}} object (\code{SingleGroupClass}). Expected to be a 1-factor model with \code{itemtype} of any combination of \code{"2PL", "graded", or "gpcm"}.}
}
\value{
A three-element \code{list}:
\itemize{
\item vector of parameter names taking the form "item.param"
\item vector of item parameter estimates
\item covariance matrix of item parameter estimates
}
}
\description{
Extract item parameter estimates and their covariance matrix from \code{\link[mirt]{mirt}}.
}
