% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{get_mirt_vcov}
\alias{get_mirt_vcov}
\title{Extract 2PL covariance matrix of item parameter estimates from \link[mirt]{mirt}.}
\usage{
get_mirt_vcov(mirt.fit.2pl)
}
\arguments{
\item{mirt.fit.2pl}{A \link[mirt]{mirt} object (\code{SingleGroupClass}) estimated for the 2PL model.}
}
\value{
The covariance matrix of 2PL item parameter estimates.
}
\description{
Extract 2PL covariance matrix of item parameter estimates from \link[mirt]{mirt}.
}
