% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_functions.R
\name{get_params}
\alias{get_params}
\title{Extract and format item parameter estimates and their covariance matrix}
\usage{
get_params(object)
}
\arguments{
\item{object}{a list of model fits (current implementation only supports objects from \code{\link[mirt]{mirt}} with \code{itemtype} of any combination of \code{"2PL", "graded", or "gpcm"} and \code{SE = TRUE})}
}
\value{
A list of item parameter estimates and covariance matrices for each group.
}
\description{
Takes a list of 1-factor model fits from \code{\link[mirt]{mirt}} (with \code{SE = TRUE}), \code{\link[lavaan]{cfa}} (with \code{std.lv = TRUE}), or \code{\link[MplusAutomation]{readModels}} (with TECH3) and formats the item parameter estimates and their covariance matrix. All \code{robustDIF} functions assume that the estimates were obtained by maximum likelihood and the covariance is asymptotically correct.

Note that the only type of fit currently supported is the \code{SingleGroupClass} of the \code{\link[mirt]{mirt}} package.

It is possible to use fits from other software with \code{robustDIF} functions, but the parameter estimates and their covariance matrices must be formatted in a particular way. For more details, see the documentation for the example dataset \code{\link[robustDIF]{rdif.eg}}.
}
\seealso{
\code{\link[robustDIF]{rdif.eg}}
}
