% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{get_starts}
\alias{get_starts}
\title{Computes the staring values for IRLS (or NR).}
\usage{
get_starts(irt.mle, par = "intercept", log = F, alpha = 0.05)
}
\arguments{
\item{irt.mle}{The output of \code{robustDIF::get_irt_mle}.}

\item{par}{Use scaling function for item intercepts or item slopes? One of \code{c("intercept", "slope")}.}

\item{log}{Logical: Use log of scaling function? Only applies if \code{par = "slope"}.}
}
\value{
A vector with elements med(Y), LTS(Y), and min Rho.
}
\description{
Computes the staring values for IRLS (or NR).
}
