% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{grad_mat}
\alias{grad_mat}
\title{Helper function to put gradient vectors into a matrix.}
\usage{
grad_mat(grad.vec)
}
\arguments{
\item{grad.vec}{A gradient vector.}
}
\value{
A matrix in which the columns are the 4*n.item-dimensional gradient vectors for each item.
}
\description{
Called internally by the \code{robustDIF} gradient functions.
}
