% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{joint_vcov}
\alias{joint_vcov}
\title{Merges the VCOV matrices of two 2PL models into a single block diagonal VCOV matrix that is conformable with the output of the \code{robustDIF} gradient functions (\code{grad_y}, \code{grad_z})}
\usage{
joint_vcov(irt.mle)
}
\arguments{
\item{irt.mle}{the output of \code{robustDIF::get_irt_mle}}
}
\value{
a block diagonal VCOV matrix in which each block is the VCOV of an item
}
\description{
Merges the VCOV matrices of two 2PL models into a single block diagonal VCOV matrix that is conformable with the output of the \code{robustDIF} gradient functions (\code{grad_y}, \code{grad_z})
}
