% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{lts}
\alias{lts}
\title{The least trimmed squares (LTS) estimate of location}
\usage{
lts(y, p = 0.5)
}
\arguments{
\item{y}{a vector of data points.}

\item{p}{the proportion of data points to trim.}
}
\value{
The LTS estimate of location of \code{y}.
}
\description{
The least trimmed squares (LTS) estimate of location
}
\seealso{
\code{\link[robustDIF]{get_starts}}
}
