% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdif_class_functions.R
\name{print.rdif}
\alias{print.rdif}
\title{S3 print method for objects of class "rdif"
Prints the estimated scaling parameter from rdif. (to be added: SE)}
\usage{
\method{print}{rdif}(object)
}
\arguments{
\item{object}{An object of class 'rdif', the saved output from the \code{rdif()} function.}
}
\value{
An object of class rdif, a string with rdif estimate value.
}
\description{
S3 print method for objects of class "rdif"
Prints the estimated scaling parameter from rdif. (to be added: SE)
}
\examples{
\dontrun {
mod <- rdif(rdif.eg)
print(mod)
}
}
