% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\docType{data}
\name{rdif.eg}
\alias{rdif.eg}
\title{An example data set with five items in two groups.}
\format{
A named list with 3 components:
\describe{
  \item{par.names}{A named \code{list} with the internal and original labels for parameters in the model.}
  \item{est}{A named \code{list}, each element containing \code{data.frame} of model parameters estimates for each group}
  \item{vcov0}{A named \code{list}, each element containing \code{data.frame} with the variance covariance matrix of the parameter estimates. The parameters appear in the order given by rdif.eg$par.names$internal.
}
}
}
\usage{
rdif.eg
}
\description{
A named list containing the maximum likelihood estimates and their estimated covariance matrix, for the 2PL IRT model fitted to 5 items in two independent groups. The first item has additive bias of .5 applied to the intercept. The groups have a mean difference of .5 standard deviations on the latent trait. The variances of the latent trait are equal in each group.
}
\keyword{datasets}
