% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{rho}
\alias{rho}
\title{The bi-square rho function.}
\usage{
rho(u, k = 1.96)
}
\arguments{
\item{u}{Can be a single value, vector, or matrix.}

\item{k}{The tuning variable. Can be a scalar or the same dimension as \code{u}.}
}
\value{
The bi-square rho function.
}
\description{
If abs(u) > k , rho(u) = k^2 / 6. Else, rho(u) =  k^2 / 6 * (1 - (1 - (u/k))^3).
}
