% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{rho}
\alias{rho}
\title{The bi-square rho function}
\usage{
rho(u, k = 1.96)
}
\arguments{
\item{u}{can be a single value, vector, or matrix}

\item{k}{the tuning parameter of the bi-square}
}
\value{
The bi-square rho function
}
\description{
The bi-square rho function
}
