% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{rho_fun}
\alias{rho_fun}
\title{The bi-square Rho function}
\usage{
rho_fun(mle, par = "intercept", log = F, alpha = 0.05, grid.width = 0.05)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{par}{character: use the scaling function for item intercepts or item slopes? One of \code{c("intercept", "slope")}.}

\item{log}{logical: use log of scaling function? Only applies if \code{par = "slope"}.}

\item{alpha}{the desired false positive rate for flagging items with DIF.}

\item{grid.width}{the width of grid points.}
}
\value{
A named list with theta values and the corresponding rho values.
}
\description{
Computes the objective function of the bi-square minimization problem in a location parameter, theta. The theta values are obtained internally by a grid search over the range of \code{\link[robustDIF]{y_fun}}. Useful for graphically diagnosing local solutions.
}
