% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{rho_grid}
\alias{rho_grid}
\title{Performs grid search for minimum value of the bi-square Rho function.}
\usage{
rho_grid(y, var_fun, alpha = 0.05, grid.width = 0.05)
}
\arguments{
\item{y}{Output of \code{y_fun}}

\item{var_fun}{Computes variance of \code{y} as a function of only the scale parameter.}

\item{alpha}{The desired false positive rate for flagging items with DIF.}

\item{grid.width}{The width of grid points.}
}
\value{
The location parameter that minimizes the bi-square Rho function.
}
\description{
Internal function used by \code{get_starts} to get staring values based on minimizing the rho-function of the bi-square. For plotting use \code{rho_fun} instead.
}
