% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{var_y}
\alias{var_y}
\title{Compute the variance of the asymptotic null distribution of IRT scaling functions.}
\usage{
var_y(mle, theta = NULL, par = "intercept", log = F)
}
\arguments{
\item{mle}{the output of \code{\link[robustDIF]{get_model_parms}}}

\item{theta}{the IRT scale parameter.}

\item{par}{character: use the scaling function for item intercepts or item slopes? One of \code{c("intercept", "slope")}.}

\item{log}{logical: use log of scaling function? Only applies if \code{par = "slope"}.}
}
\value{
A vector that contains the variance of the IRT scaling function, for each item.
}
\description{
Compute the variance of the asymptotic null distribution of IRT scaling functions.
}
\seealso{
\code{\link[robustDIF]{y_fun}}
}
