% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{var_y}
\alias{var_y}
\title{Computes the variance of asymptotic null distribution of y, for each item}
\usage{
var_y(theta, irt.mle)
}
\arguments{
\item{theta}{IRT scale parameter (mu / sigma)}

\item{irt.mle}{the output of \code{robustDIF::get_irt_mle}}
}
\value{
an n.items-dimensional vector that contains the variance of y for each item
}
\description{
Computes the variance of asymptotic null distribution of y, for each item
}
