% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{y_fun}
\alias{y_fun}
\title{Computes Y = (d1 - d0)/a1}
\usage{
y_fun(irt.mle)
}
\arguments{
\item{irt.mle}{the output of \code{robustDIF::get_irt_mle}}
}
\value{
vector of values Y
}
\description{
Computes Y = (d1 - d0)/a1
}
