% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{y_intercept}
\alias{y_intercept}
\title{The R-DIF scaling function for item intercepts.}
\usage{
y_intercept(irt.mle)
}
\arguments{
\item{irt.mle}{the output of \code{\link[robustDIF]{get_irt_pars}}}
}
\value{
A vector of Y values.
}
\description{
Computes \code{Y = (d1 - d0)/a1}. Used to test DIF on item intercepts.
}
