% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustDIF_functions.R
\name{y_intercept}
\alias{y_intercept}
\title{The scaling function for item intercepts.}
\usage{
y_intercept(irt.mle)
}
\arguments{
\item{irt.mle}{The output of \code{robustDIF::get_irt_mle}.}
}
\value{
A n.items vector of Y values.
}
\description{
Computes Y = (d1 - d0)/a1, used to test DIF on it em intercepts.
}
